Data dash no longer reloads all data on every visit
The app no longer reloads all the data everytime it is routed to the
dashboard. Instead we rely on the inmemory version of the data to be
correct.

When a metadata is opened it is reloaded, hopefully this will avoid
the worst consistency problems that could arise with this solution.